This version is an almost complete rebuild of the whole module. Almost no part of the module is untouched by the sweeping edits to try and make it more reliable and easier to understand and work with.
A brand new Particle 2.0 system - instead of the usual CPU based particle systems that are usually used on Foundry this module now uses GPU particles. The effect of that is allow for potentially tens of thousands of particles with higher end GPUs without it causing Foundry to slow to a crawl. I still suggest you keep particle counts reasonably low. Previously dust motes were created using a hacky texture based approach, now they are actually individually tracked with life spans and a high level of control. This uses the latest version of the 'pixi:particle-emitter' library.

Several new effects.

Canopy Shadows and Lighting - The ability to add shadows to a scene with the a large scale distortion effect designed to allow you to create tree shadows that sway in the breeze. There is a seperate highlight boosting function which lets you make the gaps between these shadows brighter.
Structural Shadows and Lighting - This is awesome for adding interesting interior shadows and there is a seperate control to boost the bright areas which allows for some very dramatic inside lighting.
Prism - A way to put an RGB splitting effect anywhere in your scene. This area also is used for the next feature.
Prism Glints/Gimmer - These are strong flare like effects which look great on shiny glass or metal.
Dust Motes - Completely rewritten to work much better.

Loading Screen - This can now be disabled entirely. The loading screen was designed to make it easier to hide the ugly pop-in of textures that is a currently unavoidable part of Foundry VTT loading.

The code has been organised to make it easier to see how things are interacting. Tidying, patching and improving is an ongoing process.